@@ -95,19 +95,19 @@ public enum CardAttribute implements Supplier<FilterLeaf<?>>, Comparator<Object>
     /** A card's supertypes. */
     SUPERTYPE("Supertype", (a) -> new SupertypeFilter()),
     /** Power of a creature card. */
-    POWER("Power", List.class, (a) -> new VariableNumberFilter(a, (c) -> c.power().stream().map((p) -> p.value).collect(Collectors.toList()), Card::powerVariable), (a, b) -> {
+    POWER("Power", List.class, (a) -> new VariableNumberFilter(a, (c) -> c.power().stream().map(CombatStat::value).collect(Collectors.toList()), Card::powerVariable), (a, b) -> {
         CombatStat first = CollectionUtils.convertToList(a, CombatStat.class).stream().filter(CombatStat::exists).findFirst().orElse(CombatStat.NO_COMBAT);
         CombatStat second = CollectionUtils.convertToList(b, CombatStat.class).stream().filter(CombatStat::exists).findFirst().orElse(CombatStat.NO_COMBAT);
         return first.compareTo(second);
     }),
     /** Toughness of a creature card. */
-    TOUGHNESS("Toughness", List.class, (a) -> new VariableNumberFilter(a, (c) -> c.toughness().stream().map((p) -> p.value).collect(Collectors.toList()), Card::toughnessVariable), (a, b) -> {
+    TOUGHNESS("Toughness", List.class, (a) -> new VariableNumberFilter(a, (c) -> c.toughness().stream().map(CombatStat::value).collect(Collectors.toList()), Card::toughnessVariable), (a, b) -> {
         CombatStat first = CollectionUtils.convertToList(a, CombatStat.class).stream().filter(CombatStat::exists).findFirst().orElse(CombatStat.NO_COMBAT);
         CombatStat second = CollectionUtils.convertToList(b, CombatStat.class).stream().filter(CombatStat::exists).findFirst().orElse(CombatStat.NO_COMBAT);
         return first.compareTo(second);
     }),
     /** Loyalty of a planeswalker card. */
-    LOYALTY("Loyalty", List.class, (a) -> new VariableNumberFilter(a, (Card c) -> c.loyalty().stream().map((l) -> l.value).collect(Collectors.toList()), Card::loyaltyVariable),(a, b) -> {
+    LOYALTY("Loyalty", List.class, (a) -> new VariableNumberFilter(a, (Card c) -> c.loyalty().stream().map(Loyalty::value).collect(Collectors.toList()), Card::loyaltyVariable),(a, b) -> {
         Loyalty first = CollectionUtils.convertToList(a, Loyalty.class).stream().filter(Loyalty::exists).findFirst().orElse(Loyalty.NO_LOYALTY);
         Loyalty second = CollectionUtils.convertToList(b, Loyalty.class).stream().filter(Loyalty::exists).findFirst().orElse(Loyalty.NO_LOYALTY);
         return first.compareTo(second);
