@@ -18,7 +18,6 @@ import editor.gui.editor.EditorFrame;
  *
  * @author Alec Roelke
  */
-@SuppressWarnings("serial")
 public class EditorTableTransferHandler extends EditorFrameTransferHandler
 {
     /** Editor containing the deck to import cards into. */
@@ -56,9 +55,8 @@ public class EditorTableTransferHandler extends EditorFrameTransferHandler
     @Override
     public void exportDone(JComponent source, Transferable data, int action)
     {
-        if (data instanceof EntryTransferData)
+        if (data instanceof EntryTransferData d)
         {
-            EntryTransferData d = (EntryTransferData)data;
             switch (action)
             {
             case TransferHandler.MOVE:
