@@ -16,7 +16,6 @@ import editor.util.Containment;
  *
  * @author Alec Roelke
  */
-@SuppressWarnings("serial")
 public class TypeLineFilterPanel extends FilterEditorPanel<TypeLineFilter>
 {
     /**
@@ -67,8 +66,8 @@ public class TypeLineFilterPanel extends FilterEditorPanel<TypeLineFilter>
     @Override
     public void setContents(FilterLeaf<?> filter) throws IllegalArgumentException
     {
-        if (filter instanceof TypeLineFilter)
-            setContents((TypeLineFilter)filter);
+        if (filter instanceof TypeLineFilter f)
+            setContents(f);
         else
             throw new IllegalArgumentException("Illegal type line filter " + filter.type());
     }
