@@ -12,7 +12,6 @@ import javax.swing.filechooser.FileNameExtensionFilter;
  *
  * @author Alec Roelke
  */
-@SuppressWarnings("serial")
 public class OverwriteFileChooser extends JFileChooser
 {
     /**
@@ -42,9 +41,8 @@ public class OverwriteFileChooser extends JFileChooser
     public File getSelectedFile()
     {
         File file = super.getSelectedFile();
-        if (file != null && getFileFilter() instanceof FileNameExtensionFilter)
+        if (file != null && getFileFilter() instanceof FileNameExtensionFilter filter)
         {
-            FileNameExtensionFilter filter = (FileNameExtensionFilter)getFileFilter();
             String fname = file.getAbsolutePath();
             if (Arrays.stream(filter.getExtensions()).noneMatch((ext) -> fname.endsWith("." + ext)))
                 file = new File(fname + "." + filter.getExtensions()[0]);
