@@ -6,14 +6,24 @@ import java.time.format.DateTimeFormatter;
 /**
  * This class represents an expansion set of Magic: The Gathering cards.
  *
- * @author Alec
+ * @param name name of the expansion
+ * @param block name of the block the expansion belongs to ({@link NO_BLOCK} if there is none)
+ * @param code the expansion's code
+ * @param count number of cards in the expansion
+ * @param released date the expansion was released
+ * 
+ * @author Alec Roelke
  */
-public class Expansion implements Comparable<Expansion>
+public record Expansion(String name, String block, String code, int count, LocalDate released) implements Comparable<Expansion>
 {
     /**
      * Array containing all block names.
      */
     public static String[] blocks = {};
+    /**
+     * Text to show when an expansion isn't part of a block.
+     */
+    public static final String NO_BLOCK = "<No Block>";
     /**
      * Date format used to decode expansion release dates.
      */
@@ -23,72 +33,10 @@ public class Expansion implements Comparable<Expansion>
      */
     public static Expansion[] expansions = {};
 
-    /**
-     * Name of the block this Expansion belongs to (empty if there is none).
-     */
-    public final String block;
-    /**
-     * This Expansion's code.
-     */
-    public final String code;
-    /**
-     * Number of cards in this Expansion.
-     */
-    public final int count;
-    /**
-     * This Expansion's name.
-     */
-    public final String name;
-    /**
-     * The date the expansion was released.
-     */
-    public final LocalDate releaseDate;
-
-    /**
-     * Create a new Expansion.
-     *
-     * @param name name of the new expansion
-     * @param block name of the block the new Expansion belongs to
-     * @param code code of the new Expansion
-     * @param count number of cards in the new Expansion
-     */
-    public Expansion(String name, String block, String code, int count, LocalDate date)
-    {
-        this.name = name;
-        this.block = block;
-        this.code = code;
-        this.count = count;
-        this.releaseDate = date;
-    }
-
     @Override
     public int compareTo(Expansion other)
     {
         return name.compareTo(other.name);
-    }
-
-    /**
-     * {@inheritDoc}
-     * Expansions are equal if they have the same name.
-     */
-    @Override
-    public boolean equals(Object other)
-    {
-        if (other == null)
-            return false;
-        if (other.getClass() != getClass())
-            return false;
-        if (other == this)
-            return true;
-
-        Expansion o = (Expansion)other;
-        return name.equals(o.name);
-    }
-
-    @Override
-    public int hashCode()
-    {
-        return name.hashCode();
     }
 
     /**
