@@ -30,7 +30,7 @@ public class ExpansionFilter extends SingletonOptionsFilter<Expansion>
     protected Expansion convertFromString(String str)
     {
         for (Expansion expansion : Expansion.expansions)
-            if (str.equalsIgnoreCase(expansion.name))
+            if (str.equalsIgnoreCase(expansion.name()))
                 return expansion;
         throw new IllegalArgumentException("Unknown expansion name \"" + str + "\"");
     }
@@ -47,12 +47,12 @@ public class ExpansionFilter extends SingletonOptionsFilter<Expansion>
     @Override
     protected JsonElement convertToJson(Expansion item)
     {
-        return new JsonPrimitive(item.name);
+        return new JsonPrimitive(item.name());
     }
 
     @Override
     protected Expansion convertFromJson(JsonElement item)
     {
-        return Arrays.stream(Expansion.expansions).filter((e) -> e.name.equalsIgnoreCase(item.getAsString())).findAny().get();
+        return Arrays.stream(Expansion.expansions).filter((e) -> e.name().equalsIgnoreCase(item.getAsString())).findAny().get();
     }
 }
