@@ -19,7 +19,6 @@ import editor.gui.settings.SettingsDialog;
  *
  * @author Alec Roelke
  */
-@SuppressWarnings("serial")
 public class DefaultsFilterPanel extends FilterEditorPanel<FilterLeaf<?>>
 {
     /**
@@ -41,7 +40,7 @@ public class DefaultsFilterPanel extends FilterEditorPanel<FilterLeaf<?>>
 
         categories = new HashMap<>();
 
-        var presets = SettingsDialog.settings().editor.categories.presets;
+        var presets = SettingsDialog.settings().editor().categories().presets();
         String[] names = presets.stream().map(CategorySpec::getName).toArray(String[]::new);
         for (int i = 0; i < presets.size(); i++)
             categories.put(names[i], presets.get(i));
