@@ -121,38 +121,38 @@ public class SettingsBuilder
      */
     public SettingsBuilder copy(Settings original)
     {
-        inventorySource = original.inventory.source;
-        inventoryFile = original.inventory.file;
-        inventoryVersionFile = original.inventory.versionFile;
-        inventoryVersion = original.inventory.version;
-        inventoryLocation = original.inventory.location;
-        inventoryScans = original.inventory.scans;
-        imageSource = original.inventory.imageSource;
-        enableImageLimit = original.inventory.imageLimitEnable;
-        imageLimit = original.inventory.imageLimit;
-        inventoryTags = original.inventory.tags;
-        inventoryUpdate = original.inventory.update;
-        inventoryWarn = original.inventory.warn;
-        inventoryColumns = original.inventory.columns;
-        inventoryBackground = original.inventory.background;
-        inventoryStripe = original.inventory.stripe;
-        recentsCount = original.editor.recents.count;
-        recentsFiles = original.editor.recents.files;
-        presetCategories = new ArrayList<>(original.editor.categories.presets.stream().map(CategorySpec::new).collect(Collectors.toList()));
-        categoryRows = original.editor.categories.rows;
-        explicits = original.editor.categories.explicits;
-        editorColumns = original.editor.columns;
-        editorStripe = original.editor.stripe;
-        handSize = original.editor.hand.size;
-        handRounding = original.editor.hand.rounding;
-        handBackground = original.editor.hand.background;
-        searchForCommander = original.editor.legality.searchForCommander;
-        main = original.editor.legality.main;
-        all = original.editor.legality.all;
-        list = original.editor.legality.list;
-        sideboard = original.editor.legality.sideboard;
-        manaValue = original.editor.manaValue;
-        cwd = original.cwd;
+        inventorySource = original.inventory().source();
+        inventoryFile = original.inventory().file();
+        inventoryVersionFile = original.inventory().versionFile();
+        inventoryVersion = original.inventory().version();
+        inventoryLocation = original.inventory().location();
+        inventoryScans = original.inventory().scans();
+        imageSource = original.inventory().imageSource();
+        enableImageLimit = original.inventory().imageLimitEnable();
+        imageLimit = original.inventory().imageLimit();
+        inventoryTags = original.inventory().tags();
+        inventoryUpdate = original.inventory().update();
+        inventoryWarn = original.inventory().warn();
+        inventoryColumns = original.inventory().columns();
+        inventoryBackground = original.inventory().background();
+        inventoryStripe = original.inventory().stripe();
+        recentsCount = original.editor().recents().count();
+        recentsFiles = original.editor().recents().files();
+        presetCategories = new ArrayList<>(original.editor().categories().presets().stream().map(CategorySpec::new).collect(Collectors.toList()));
+        categoryRows = original.editor().categories().rows();
+        explicits = original.editor().categories().explicits();
+        editorColumns = original.editor().columns();
+        editorStripe = original.editor().stripe();
+        handSize = original.editor().hand().size();
+        handRounding = original.editor().hand().rounding();
+        handBackground = original.editor().hand().background();
+        searchForCommander = original.editor().legality().searchForCommander();
+        main = original.editor().legality().main();
+        all = original.editor().legality().all();
+        list = original.editor().legality().list();
+        sideboard = original.editor().legality().sideboard();
+        manaValue = original.editor().manaValue();
+        cwd = original.cwd();
 
         return this;
     }
